edit su

+++ scripts/build.hwloc.sh
@@ -7,8 +7,10 @@ mkdir -p deps/include
 mkdir -p deps/lib
 
 mkdir -p build && cd build
+if [ -f hwloc-${HWLOC_VERSION}.tar.gz ] ; then
 
 wget https://download.open-mpi.org/release/hwloc/v2.4/hwloc-${HWLOC_VERSION}.tar.gz -O hwloc-${HWLOC_VERSION}.tar.gz
+fi
 tar -xzf hwloc-${HWLOC_VERSION}.tar.gz
 
 cd hwloc-${HWLOC_VERSION}
