+++ scripts/build.uv.sh
@@ -1,16 +1,13 @@
 #!/bin/bash -e
-
 UV_VERSION="1.40.0"
-
 mkdir -p deps
 mkdir -p deps/include
 mkdir -p deps/lib
-
 mkdir -p build && cd build
-
+if [ -f v${UV_VERSION}.tar.gz ] ; then
 wget https://github.com/libuv/libuv/archive/v${UV_VERSION}.tar.gz -O v${UV_VERSION}.tar.gz
+fi
 tar -xzf v${UV_VERSION}.tar.gz
-
 cd libuv-${UV_VERSION}
 sh autogen.sh
 ./configure --disable-shared
